pyproject.toml
[project]
name = "investment-pipeline-project-portfolio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "delta-spark>=4.1.0",
    "google-cloud-storage>=3.10.1",
    "pyarrow>=23.0.1",
    "pyspark>=4.1.1",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "yfinance>=1.2.0",
]
